move code to floss.results module
