+++ floss/main.py
@@ -25,7 +25,7 @@ import floss.identification_manager as im
 from floss.const import MAX_FILE_SIZE, DEFAULT_MIN_LENGTH, SUPPORTED_FILE_MAGIC
 from floss.utils import hex, get_vivisect_meta_info
 from floss.version import __version__
-from floss.render.result_document import (
+from floss.results import (
     Metadata,
     AddressType,
     StackString,
