+++ scripts/render-r2-import-script.py
@@ -24,7 +24,7 @@ import base64
 import logging
 import argparse
 
-from floss.render.result_document import AddressType, ResultDocument
+from floss.results import AddressType, ResultDocument
 
 logger = logging.getLogger("floss.render-r2-import-script")
