+++ scripts/render-x64dbg-database.py
@@ -24,7 +24,7 @@ import logging
 import os.path
 import argparse
 
-from floss.render.result_document import AddressType, ResultDocument
+from floss.results import AddressType, ResultDocument
 
 logger = logging.getLogger("floss.render-x64dbg-import-script")
